services/core-api/pyproject.toml
[project]
name = "neft-core-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn==0.30.6",
  "pydantic==2.9.2",
  "pydantic-settings==2.7.0",
  "SQLAlchemy==2.0.32",
  "psycopg[binary]==3.2.1",
  "alembic==1.13.2",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "redis==5.0.7",
  "httpx==0.27.2",
  "ruff==0.6.9",
  "pytest==8.3.3"
]
